==> src/16_get_Vietnam.ipynb <==
import numpy as np
import pandas as pd
from tqdm import tqdm
import cudf

tqdm.pandas()

dfall = pd.read_parquet("checkpoints/preprocessed_2023-08-28T11-09-39.parquet")

vn = dfall[dfall["geocode_country_name"] == "Vietnam"]
vn = vn[vn.fundraisingDate > "2020-01-01"]

vn.project_id.nunique()

dfall.head()
