==> src/21_build_graph_vn.ipynb <==
DEVICES = "0,1,2"

import os

os.environ["CUDA_VISIBLE_DEVICES"] = DEVICES

import cupy as cp
import pandas as pd
from tqdm import tqdm
import cudf
import networkx as nx


tqdm.pandas()

# Import raw data
First, read data in `.jsonl` file format as a pandas data frame
Then store the dataframe in `.parquet` format for easy access later

%%script false --no-raise-error

df = pd.read_json("../fulldata/kiva_activity_2023-08-28T11-09-39.jsonl", lines=True)
df = pd.json_normalize(df["loan"], sep='_')

%%script false --no-raise-error

df["loanAmount"] = df["loanAmount"].astype(float)
df["loanFundraisingInfo_fundedAmount"] = df["loanFundraisingInfo_fundedAmount"].astype(float)
df["raisedDate"] = pd.to_datetime(df["raisedDate"])
df["fundraisingDate"] = pd.to_datetime(df["fundraisingDate"])
df["geocode_country_name"] = df["geocode_country_name"].astype("category")
df["sector_id"] = df["sector_id"].astype(int)
df["sector_name"] = df["sector_name"].astype("category")
df["activity_id"] = df["activity_id"].astype(int)
df["activity_name"] = df["activity_name"].astype("category")

%%script false --no-raise-error
df.to_parquet("../fulldata/kiva_activity_2023-08-28T11-09-39.parquet")

ds = cudf.read_parquet("../fulldata/kiva_activity_2023-08-28T11-09-39.parquet")

ds.dropna(axis=0, how="all", inplace=True)
ds.tail()

# Filter data for Vietnam only

Filtering, only take `Vietnam` into account
Why? Because there are a lot of rows and we try to localize the task

ds["geocode_country_name"].value_counts()["Vietnam"]

ds = ds[ds["geocode_country_name"] == "Vietnam"]

Try to limit the timeline, because I am not yet comfortable to work with large data

ds = ds[ds.fundraisingDate > "2020-01-01"]
"the number of Loans (might duplicated) under investigation is", len(ds)

ds.tail()

# Basic process

ds.rename(columns={"id": "project_id", "name": "project_name"}, inplace=True)

ds.drop(columns=["tags", "lendingActions_values"]).info()

ds["sector_name"] = ds["sector_name"].astype("category")
ds["geocode_country_name"] = ds["geocode_country_name"].astype("category")
ds["activity_name"] = ds["activity_name"].astype("category")

# Preprocessing

## Remove duplicated Projects

There are Projects which have a same `project_id` but different `fundedAmount`
It might because the query time is different
Here, only keep records which have the highest `fundedAmount`

ds.loc[[9628, 1366545]]

temp = ds.groupby("project_id", group_keys=False)[["loanFundraisingInfo_fundedAmount"]].idxmax()
iloc = temp["loanFundraisingInfo_fundedAmount"].values  # NOTE: just iloc, not loc
ds = ds.iloc[iloc]
del iloc
del temp
ds.loc[[9628, 1366545]]  # see, only keep the one with higher fundedAmount

assert 0 == len(ds[ds.duplicated(subset=["project_id"], keep=False)].sort_values(by=["project_id"]))  # no duplicated

"the number of Projects (no duplicated) under investigation is", len(ds)

## create `Lender-Project-Tag` df

ads = ds.explode("lendingActions_values")
len(ads)

# drop some loans that has no lender
ads.dropna(subset=["lendingActions_values"], inplace=True)

# dict processing cant be done in cuDF, so convert to pandas
adf = ads.to_pandas()

adf["lender_id"] = adf.progress_apply(lambda x: x["lendingActions_values"]["lender"]["id"], axis=1).astype(int)
adf["lender_name"] = adf.progress_apply(lambda x: x["lendingActions_values"]["lender"]["name"], axis=1)
adf["lender_publicId"] = adf.progress_apply(lambda x: x["lendingActions_values"]["lender"]["publicId"], axis=1)
adf["loan_shareAmount"] = adf.progress_apply(lambda x: x["lendingActions_values"]["shareAmount"], axis=1).astype(float)
adf["loan_date"] = pd.to_datetime(
    adf.progress_apply(lambda x: x["lendingActions_values"]["latestSharePurchaseDate"], axis=1)
)

# cuDF do not work with timezone yet
adf["loan_date"] = adf["loan_date"].dt.tz_localize(None)

adf.drop(["lendingActions_values"], axis=1, inplace=True)

ads = cudf.from_pandas(adf)
del adf

assert 0 == ads.duplicated(subset=["project_id", "lender_id", "loan_shareAmount", "loan_date"]).sum()

ads["loan_id"] = cp.arange(len(ads["lender_id"])) + 1

explode `tags`

ads = ads.explode("tags")

## Tag preprocessing

# there are many Loans that do not have tags
ads[ads["tags"].isna()].project_id.unique().count(), "~", ads[
    ads["tags"].isna()
].project_id.unique().count() / ads.project_id.unique().count() * 100, "percent"

# create a tag call `empty`
# ads.dropna(subset=["tags"], inplace=True)
ads[["tags"]].fillna("empty", inplace=True)

Remove some tags 
The folowing tags should be remove, because it isn't visible to Users:  
- `user_favorite`
- `user_like`
- `volunteer_like`
- `volunteer_pick`

If a project **only** have those tags, change all those tags into `empty`, then remove duplicate again.  
If a project have other tags rather than those tags, just drop those tags

(ads["tags"] == "user_favorite").sum(), (ads["tags"] == "user_like").sum(), (ads["tags"] == "volunteer_like").sum(), (
    ads["tags"] == "volunteer_pick"
).sum()

ads.duplicated().sum()  # NOTE: only work with small dataset

ads["tags"] = ads.tags.replace(["user_favorite", "user_like", "volunteer_like", "volunteer_pick"], ["removetag"] * 4)

ads.drop_duplicates(inplace=True)  # NOTE: only work with small dataset

# count tags by loans
# NOTE: this method only support small size data, consider using `transform` instead
# NOTE: also note that, `transform('nunique')` might not work with cudf yet
one_tag_loans = ads.groupby("project_id").tags.nunique(dropna=False) == 1
one_tag_loans = one_tag_loans[one_tag_loans]

should_change_tag = (ads["project_id"].isin(one_tag_loans.index)) & (ads["tags"] == "removetag")
ads[should_change_tag]

ads[should_change_tag].tags = "empty"

ads = ads[ads.tags != "removetag"]

ads["tags"] = ads["tags"].astype("category").cat.as_ordered()
ads["tags"].dtype

# what is the portion of Loans that have no tags?
ads[ads.tags == "empty"].project_id.nunique(), ads.project_id.nunique(), ads[
    ads.tags == "empty"
].project_id.nunique() / ads.project_id.nunique()

There are $\sim 8\%$ of no-tag Loans, just remove them

ads = ads[ads.tags != "empty"]

## Remove anonymous Lenders

ads.head()

print("anynomous Lenders in Vietnam")
anons = ads.lender_publicId.str.startswith("anon")
ads[anons][["lender_id", "lender_name", "lender_publicId"]].drop_duplicates()

# drop those anon
ads = ads[~anons]

ads.drop_duplicates(inplace=True)

ads.reset_index(inplace=True)

ads["sector_name"] = ads["sector_name"].astype(str)
ads["geocode_country_name"] = ads["geocode_country_name"].astype(str)
ads["activity_name"] = ads["activity_name"].astype(str)
ads["tags"] = ads["tags"].astype(str)
ads.to_parquet("checkpoints/vn_since_20200101.parquet")

ads = cudf.read_parquet("checkpoints/vn_since_20200101.parquet")
ads["sector_name"] = ads["sector_name"].astype("category")
ads["geocode_country_name"] = ads["geocode_country_name"].astype("category")
ads["activity_name"] = ads["activity_name"].astype("category")
ads["tags"] = ads["tags"].astype("category")

# Stats for Vietnam

"number of Loans for Vietnam", ads.project_id.unique().count()

"number of Lenders for Vietnam", ads.lender_id.unique().count()

# Contruct a Graph

The idea is construct a graph with following node type
- `Lender`
- `Loan`
- `Tag`

With following relationships
- `Lender`s can `LEND` to `Loan`s
- `Loan`s can be `TAGGED_WITH` `Tag`s

Lenders have properties
- `id`
- `name`
- `publicId`

Loan have properties
- `id`
- `name`
- `loanAmount`
- `fundedAmount`
- `postDate`
- `raisedDate`

`Tag` have properties:
- `name`

LEND's properties
- `shareAmount`
- `date`

TAGGED_WITH have no properties

## create `Tag` nodes

# create those df
ds_tags = ads[["tags"]].drop_duplicates().dropna()
ds_tags[":LABEL"] = "Tag"
# save in neo4j style
ds_tags.rename(columns={"tags": "name:ID"}).to_csv("../data/gen/tags.csv", index=False)
print(len(ds_tags))
ds_tags.head()

## create `Project` nodes

ds_loan = ads.drop(
    [
        "tags",
        "lendingActions_totalCount",
        "lender_id",
        "lender_name",
        "lender_publicId",
        "loan_shareAmount",
        "loan_date",
        "loan_id",
    ],
    axis=1,
).drop_duplicates()
len(ds_loan)

ds_loan[":LABEL"] = "Project"
ds_loan.rename(columns={"project_id": "id:ID(Project-ID)"}).to_csv("../data/gen/projects.csv", index=False)
ds_loan.head(5)

## create `Lender` nodes

ds_lender = ads[["lender_id", "lender_name", "lender_publicId"]].drop_duplicates()
ds_lender.tail(2)

ds_lender[ds_lender.duplicated(subset=["lender_id"], keep=False)].sort_values(by=["lender_id"])

duplicated_lender_id = ds_lender[ds_lender.duplicated(subset=["lender_id"])]["lender_id"]
should_remove = ds_lender[(ds_lender["lender_id"].isin(duplicated_lender_id)) & (ds_lender["lender_publicId"].isna())]
ds_lender.drop(should_remove.index, axis=0, inplace=True)

# drop duplicated_lender who publicId is None
duplicated_lender_id = ds_lender[ds_lender.duplicated(subset=["lender_id"])]["lender_id"]
should_remove = ds_lender[(ds_lender["lender_id"].isin(duplicated_lender_id)) & (ds_lender["lender_publicId"].isna())]
ds_lender.drop(should_remove.index, axis=0, inplace=True)
# still duplicate, might be because user change name and publicId. Just remove duplicates here.
ds_lender.drop_duplicates(subset="lender_id", inplace=True)
del duplicated_lender_id
del should_remove
# display the duplicated
ds_lender[ds_lender.duplicated(subset=["lender_id"], keep=False)]

ds_lender[":LABEL"] = "Lender"
ds_lender.rename(columns={"lender_id": "id:ID(Lender-ID)"}).to_csv("../data/gen/lenders.csv", index=False)
print(len(ds_lender))
ds_lender.head()

## Create `TAGGED_WITH` relationship between `Loan` and `Tags`

ds_loan_tags = ads[["project_id", "tags"]].dropna()
ds_loan_tags.isna().sum()

ds_loan_tags.drop_duplicates(inplace=True)
ds_loan_tags.duplicated().sum()

"the number of projects is ", len(ds_loan_tags["project_id"].drop_duplicates())

"the number of loan-tag relationships is", len(ds_loan_tags)

ds_loan_tags["tags"].value_counts()

ds_loan_tags[":TYPE"] = "TAGGED_WITH"
ds_loan_tags.rename(columns={"id": ":START_ID(Project-ID)", "tags": ":END_ID"}).to_csv(
    "../data/gen/project_tags.csv", index=False
)
print(len(ds_loan_tags))
ds_loan_tags.head()

## create `LEND` relationship between `Lender` and `Loan`

ds_lender_loan = ads[["project_id", "lender_id", "lender_publicId", "loan_shareAmount", "loan_date"]]
ds_lender_loan.dropna(inplace=True)
ds_lender_loan.tail(5)

ds_lender_loan.drop_duplicates(inplace=True)  # duplicated cause by exploding the tags

assert len(ds_lender_loan) == ads.loan_id.nunique()

ds_lender_loan[":TYPE"] = "LEND"
ds_lender_loan.rename(columns={"lender_id": ":START_ID(Lender-ID)", "project_id": ":END_ID(Project-ID)"}).to_csv(
    "../data/gen/lender_project.csv", index=False
)
print(len(ds_lender_loan))
ds_lender_loan.head()

## [Optional] Manually creat the `INTEREST` relationship

It is now easy to create new relationship in frameworks like networkx, cugraph and gephi. So, create them here

lender_tag_ds = ads[["project_id", "tags", "lender_id", "lender_publicId", "loan_shareAmount", "loan_date"]]
lender_tag_ds.tail(3)

assert 0 == lender_tag_ds.duplicated().sum()

lender_tag_ds = lender_tag_ds.groupby(["lender_id", "tags"]).agg(
    {"lender_publicId": "first", "loan_shareAmount": "sum", "loan_date": "count"}
)
lender_tag_ds.reset_index(inplace=True)
lender_tag_ds

lender_tag_ds.rename(columns={"loan_shareAmount": "loan_shareAmount_sum", "loan_date": "loan_date_count"}, inplace=True)
lender_tag_ds.sort_values(by=["loan_shareAmount_sum"], ascending=False).head()

# Generate `cuGraph` network

# verify that we could use lender_publicId instead of lender_id
assert ds_lender.lender_id.nunique() == ds_lender.lender_publicId.nunique()

lender_tag_ds.head(1)

# create a graph using nodes from df_tags, df_loan, df_lender, edges from df_loan_tags and df_lender_loan
# NOTE: this is a 3-partite graph

G = nx.DiGraph()

# first, add tags node from df_tags
G.add_nodes_from(ds_tags.to_pandas()["tags"], type="Tag")

# NOTE: here we do not use lender_id as node, because it may duplicated with project_id
# instead, we use lender_publicId as node

for row in ds_lender.to_pandas().itertuples():
    G.add_node(row.lender_publicId, type="Lender", lender_id=row.lender_id)

for row in ds_loan.to_pandas().itertuples():
    G.add_node(
        row.project_id,
        type="Project",
        # postDate=row.fundraisingDate,
        # raisedDate=row.raisedDate,
        loanAmount=row.loanAmount,
        fundedAmount=row.loanFundraisingInfo_fundedAmount,
        sector=row.sector_name,
        country=row.geocode_country_name,
        activity=row.activity_name,
    )


for row in ds_loan_tags.to_pandas().itertuples():
    G.add_edge(row.project_id, row.tags, label="TAGGED_WITH")

for row in ds_lender_loan.to_pandas().itertuples():
    # G.add_edge(row.lender_id, row.project_id, type="LEND", shareAmount=row.loan_shareAmount, date=row.loan_date)
    G.add_edge(row.lender_publicId, row.project_id, label="LEND", shareAmount=row.loan_shareAmount)

for row in lender_tag_ds.to_pandas().itertuples():
    G.add_edge(
        row.lender_publicId, row.tags, label="INTEREST", shareAmount=row.loan_shareAmount_sum, count=row.loan_date_count
    )

# print some basic information of G
# count number of nodes for each type of node
node_types = set(nx.get_node_attributes(G, "type").values())
for node_type in node_types:
    count = sum(1 for n, d in G.nodes(data=True) if d["type"] == node_type)
    print(f"Number of nodes for {node_type}: {count}")

# count number of edges for each type of edge
edge_types = set(nx.get_edge_attributes(G, "label").values())
for edge_type in edge_types:
    count = sum(1 for u, v, d in G.edges(data=True) if d["label"] == edge_type)
    print(f"Number of edges for {edge_type}: {count}")

# save G in a way that Gephi can read
nx.write_gexf(G, "../data/gen/vn_since_20200101.gexf")

# ALL following is DEPRECATED

raise SystemExit("expectedlly ended")

## Manually create the `COMMON_LOAN` relationship between `Tag`s

ds_loan_tags.drop(":TYPE", axis=1, inplace=True)
ds_loan_tags.head()

self_merged = ds_loan_tags.merge(ds_loan_tags, on="id")
self_merged.head()

filtered = self_merged[self_merged["tags_x"] > self_merged["tags_y"]]
filtered.head()

ds_common_loans = filtered.groupby(["tags_x", "tags_y"])["id"].nunique().reset_index()
ds_common_loans.rename({"id": "commond_loan_count"}, axis=1, inplace=True)
ds_common_loans

ds_common_loans[":TYPE"] = "COMMON_LOAN"
ds_common_loans.rename(columns={"tags_x": ":START_ID", "tags_y": ":END_ID", "commond_loan_count": "weight"}).to_csv(
    "../data/gen/tag_tag_common_loan.csv", index=False
)

## Manually create the `SHARES_LOANS` relationship

ds_lender_loan = ads[["id", "lender_id", "shareAmount", "date"]]
ds_lender_loan.dropna(inplace=True)
ds_lender_loan.drop_duplicates(inplace=True)  # duplicated cause by exploding the tags
ds_lender_loan.drop(["shareAmount", "date"], axis=1, inplace=True)
ds_lender_loan.tail(2)

ds_lender_loan.to_feather("ds_lender_loan.feather")

ds_lender_loan = cudf.read_feather("ds_lender_loan.feather")

ds_lender_loan.info()

ds_lender_loan.lender_id.nunique()

ds_lender_loan.id.unique()

ds_lender_loan.id.nunique()

"expected number of row of self_merge is", pow(
    ds_lender_loan.lender_id.nunique() / ds_lender_loan.id.nunique(), 2
) * ds_lender_loan.id.nunique()

self_merged = ds_lender_loan.merge(ds_lender_loan, on="id")
del ds_lender_loan
self_merged.head(3)

self_merged = self_merged[self_merged["lender_id_x"] > self_merged["lender_id_y"]]

shares_loan = self_merged.groupby(["lender_id_x", "lender_id_y"]).nunique().reset_index()
shares_loan.rename(columns={"id": "number_common_loans"}, inplace=True)
del self_merged
shares_loan.head()

len(shares_loan)

shares_loan.rename(
    columns={"lender_id_x": ":START_ID(Lender-ID)", "lender_id_y": ":END_ID(Lender-ID)", "id": "weight"}, inplace=True
)
shares_loan[":TYPE"] = "SHARES_LOAN"
shares_loan.to_csv("../data/gen/lender_lender_share_loan.csv", index=False)
del shares_loan

## Manually create the `SHARE_TAGS` relationship

lender_tag_ds.drop(["weight", ":TYPE"], axis=1, inplace=True)
lender_tag_ds.head()

lender_tag_ds.info()

import gc

gc.collect()

ads.memory_usage().sum() / pow(2, 30)

lender_tag_ds.memory_usage().sum() / pow(2, 30)

lender_tag_ds["tags"] = lender_tag_ds["tags"].astype(str)
lender_tag_ds.to_feather("checkpoints/lender_tag_ds.feather")

lender_tag_ds = cudf.read_feather("gen/lender_tag_ds.feather")
lender_tag_ds["tags"] = lender_tag_ds["tags"].astype("category")

# %%script false --no-raise-error

from sqlalchemy import create_engine, text

# engine = create_engine("sqlite:///lender_tag.sqlite", echo=False)
engine = create_engine("postgresql+psycopg2://postgres:postgres@localhost:32772/datdb", echo=False)

# %%script false --no-raise-error

lender_tag_ds.to_pandas().to_sql(name="lender_tag", con=engine, index=False)

use the following query

```sql
CREATE TABLE merged AS
select 
l1.lender_id as lender_id_x, 
l1.tags as tag_x,
l2.lender_id as lender_id_y,
l2.tags as tag_y
from lender_tag as l1, lender_tag as l2 
where l1.tags = l2.tags
and l1.lender_id > l2.lender_id
```

Since I've kept `empty` tag, must remove it before doing further queries

```sql
alter table merged drop column tag_y;
alter table merged rename column tag_x to tag;
CREATE INDEX tag_idx on merged using hash(tag);
CREATE INDEX lender_x_y_idx on merged(lender_id_x, lender_id, y);
delete from merged where tag = 'empty';
```

with engine.connect() as conn:
    share_tags_ds = pd.read_sql(
        text("select lender_id_x, lender_id_y, tag_x as tag from merged where tag_x != 'empty'"), conn
    )
share_tags_ds.to_parquet("checkpoints/share_tags_ds_vietnam.parquet")
len(share_tags_ds)
