==> src/52_stable_lender_keep.ipynb <==
import pandas as pd
from typing import Dict, Set
import seaborn as sns
import matplotlib.pyplot as plt

from tqdm import tqdm

tqdm.pandas()

base on the last draw, we will filter data from 2019 to 2023, and keep only user who actively invested in all the years from 2019 to 2023.

df = pd.read_parquet("../data/gen/preprocessed_2023-08-28T11-09-39.parquet")

df = df[df["fundraisingDate"] >= "2019-01-01"]

each_year = df.groupby(df["fundraisingDate"].dt.year).agg({"lender_id": set})
each_year["lender_count"] = each_year["lender_id"].apply(len)
each_year.reset_index(inplace=True)
each_year.tail()

union = set()
for s in each_year["lender_id"]:
    union = union.union(s)

len(union)

intersec = union
for s in each_year["lender_id"]:
    intersec = intersec.intersection(s)

len(intersec)

len(intersec) / len(union)

active_lender = intersec

# how many project that these active lenders have funded?
df[df["lender_id"].isin(active_lender)]["project_id"].nunique()

# how many total projects?
df["project_id"].nunique()

# how many loans that these active lenders have funded?
df[df["lender_id"].isin(active_lender)]["loan_id"].nunique()

# how many loans in total?
df["loan_id"].nunique()

7368409 / 11128507

# keep only loan that has active lenders
df = df[df["lender_id"].isin(active_lender)]

# save it, prepare to lunch biLouvian again
df.to_parquet("../data/gen/preprocessed_2023-08-28T11-09-39_from_2019-01-01_activelender.parquet")
